+++ tests/matrix_test.cpp
@@ -18,7 +18,7 @@ BOOST_AUTO_TEST_SUITE(random_matrix)
 BOOST_AUTO_TEST_CASE(random_matrix_2x2) {
   AdjacencyMatrix matrix = AdjacencyMatrix({{FLT_MAX, 80}, {91, FLT_MAX}});
 
-  float asmt = 171;
+  double asmt = 171;
   std::pair<int, int> edge = {1, 2};
   BOOST_TEST(matrix.GetBottomLineEvaluation() == asmt);
   BOOST_TEST(matrix.GetSelectedEdge().first == edge.first - 1);
@@ -29,7 +29,7 @@ BOOST_AUTO_TEST_CASE(random_matrix_3x3) {
   AdjacencyMatrix matrix = AdjacencyMatrix(
       {{FLT_MAX, 80, 98}, {91, FLT_MAX, 59}, {16, 21, FLT_MAX}});
 
-  float asmt = 155;
+  double asmt = 155;
   std::pair<int, int> edge = {2, 3};
   BOOST_TEST(matrix.GetBottomLineEvaluation() == asmt);
   BOOST_TEST(matrix.GetSelectedEdge().first == edge.first - 1);
@@ -42,7 +42,7 @@ BOOST_AUTO_TEST_CASE(random_matrix_4x4) {
                                             {13, 75, FLT_MAX, 63},
                                             {88, 24, 58, FLT_MAX}});
 
-  float asmt = 138;
+  double asmt = 138;
   std::pair<int, int> edge = {4, 2};
   BOOST_TEST(matrix.GetBottomLineEvaluation() == asmt);
   BOOST_TEST(matrix.GetSelectedEdge().first == edge.first - 1);
@@ -56,7 +56,7 @@ BOOST_AUTO_TEST_CASE(random_matrix_5x5) {
                                             {4, 78, 32, FLT_MAX, 32},
                                             {57, 86, 83, 24, FLT_MAX}});
 
-  float asmt = 146;
+  double asmt = 146;
   std::pair<int, int> edge = {3, 2};
   BOOST_TEST(matrix.GetBottomLineEvaluation() == asmt);
   BOOST_TEST(matrix.GetSelectedEdge().first == edge.first - 1);
@@ -73,7 +73,7 @@ BOOST_AUTO_TEST_CASE(symmetric_matrix_4x4) {
                                             {99, 81, FLT_MAX, 55},
                                             {64, 25, 55, FLT_MAX}});
 
-  float asmt = 146;
+  double asmt = 146;
   std::pair<int, int> edge = {1, 2};
   BOOST_TEST(matrix.GetBottomLineEvaluation() == asmt);
   BOOST_TEST(matrix.GetSelectedEdge().first == edge.first - 1);
@@ -88,7 +88,7 @@ BOOST_AUTO_TEST_CASE(symmetric_matrix_6x6) {
                                             {18, 16, 19, 35, FLT_MAX, 15},
                                             {12, 54, 69, 73, 15, FLT_MAX}});
 
-  float asmt = 118;
+  double asmt = 118;
   std::pair<int, int> edge = {2, 3};
   BOOST_TEST(matrix.GetBottomLineEvaluation() == asmt);
   BOOST_TEST(matrix.GetSelectedEdge().first == edge.first - 1);
@@ -108,7 +108,7 @@ BOOST_AUTO_TEST_CASE(symmetric_matrix_10x10) {
                        {492, 272, 20, 414, 340, 90, 105, 369, FLT_MAX, 377},
                        {62, 127, 446, 330, 406, 189, 379, 422, 377, FLT_MAX}});
 
-  float asmt = 659;
+  double asmt = 659;
   std::pair<int, int> edge = {5, 4};
   BOOST_TEST(matrix.GetBottomLineEvaluation() == asmt);
   BOOST_TEST(matrix.GetSelectedEdge().first == edge.first - 1);
@@ -125,7 +125,7 @@ BOOST_AUTO_TEST_CASE(symm_obstacle_wise_matrix_4x4) {
                                             {99, 81, FLT_MAX, 55},
                                             {64, FLT_MAX, 55, FLT_MAX}});
 
-  float asmt = 146;
+  double asmt = 146;
   std::pair<int, int> edge = {1, 2};
   BOOST_TEST(matrix.GetBottomLineEvaluation() == asmt);
   BOOST_TEST(matrix.GetSelectedEdge().first == edge.first - 1);
@@ -141,7 +141,7 @@ BOOST_AUTO_TEST_CASE(symm_obstacle_wise_matrix_6x6) {
                        {18, 16, 19, 35, FLT_MAX, 15},
                        {FLT_MAX, 54, 69, 73, 15, FLT_MAX}});
 
-  float asmt = 146;
+  double asmt = 146;
   std::pair<int, int> edge = {5, 6};
   BOOST_TEST(matrix.GetBottomLineEvaluation() == asmt);
   BOOST_TEST(matrix.GetSelectedEdge().first == edge.first - 1);
@@ -161,7 +161,7 @@ BOOST_AUTO_TEST_CASE(symmetric_obstacle_wise_matrix_10x10) {
        {492, 272, 20, FLT_MAX, FLT_MAX, 90, 105, 369, FLT_MAX, 377},
        {62, 127, 446, 330, 406, 189, 379, 422, 377, FLT_MAX}});
 
-  float asmt = 708;
+  double asmt = 708;
   std::pair<int, int> edge = {10, 1};
   BOOST_TEST(matrix.GetBottomLineEvaluation() == asmt);
   BOOST_TEST(matrix.GetSelectedEdge().first == edge.first - 1);
